% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_timepoints.r
\name{back_timepoints}
\alias{back_timepoints}
\title{Back Timepoints}
\usage{
back_timepoints(fy, quart, num)
}
\arguments{
\item{fy}{fiscal year of most recent timepoint}

\item{quart}{quarter of most recent timepoint}

\item{num}{number of quarters to go back}
}
\value{
dataframe with list of fy & quart
}
\description{
Create a reference data table with all the backing time points
  starting from the given timepoint and stepping backwards.
}
\note{
Put everything in terms of sum of quarters 
  mapping q1 to 0 and q4 to 3 because 2000q4 != 2005
  e.g 2000Q1 => 8000, 2001Q4 => 8007
}
