% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_all_tex.r
\name{build_all_tex}
\alias{build_all_tex}
\title{Build All Tex}
\usage{
build_all_tex(rate_df, category_df, memo_type, output_dir)
}
\arguments{
\item{rate_df}{dataframe of rate data to be plotted}

\item{category_df}{dataframe of category data to be plotted}

\item{memo_type}{"hbpc" | "clc" informing which tempalte to use.}
}
\description{
Build all the LaTeX report outputs for all ids in the input data
}
