% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_category_sums.r
\name{calc_category_sums}
\alias{calc_category_sums}
\title{Calculate Category Sums}
\usage{
calc_category_sums(flt_data, id_cols, numer_categories, preserve_cols)
}
\arguments{
\item{flt_data}{dataframe of filtered data}

\item{id_cols}{names of columns to use as id}

\item{numer_categories}{list of numerator groupings each containing a vector of names of column names which belong to the numerator group.}

\item{preserve_cols}{names of columns to be preserved for subsquent use in faceting for further processing.}
}
\value{
dataframe of category sums
}
\description{
Sum the counts for categories of numerators.  This function assumes there is only a single denominator.
}
