% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_plot.r
\name{odd_palette}
\alias{odd_palette}
\alias{category_extra_colnames}
\title{Odd Palette}
\usage{
odd_palette(n)

category_extra_colnames(cnames)
}
\description{
set diff of column names in plot data with expected names
}
\section{Functions}{
\itemize{
\item \code{odd_palette}: Plot

\item \code{category_extra_colnames}: Plot
}}

