% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_plot.r
\name{category_plot}
\alias{category_plot}
\title{Category Plot}
\usage{
category_plot(plot_data, plot_title, y_label, cat_labels)
}
\arguments{
\item{plot_data}{category plot dataframe}

\item{plot_title}{string for plot title}

\item{y_label}{string for y-axis label}

\item{cat_labels}{legend text for the numerator and misses values.}

\item{line_label}{legend text for line that represents total admissions}
}
\description{
Create a category plot figure.
}
