% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_plot_data.r
\name{category_plot_data}
\alias{category_plot_data}
\title{Category Plot Data}
\usage{
category_plot_data(input_data)
}
\arguments{
\item{input_data}{dataframe with "identifier", "timepoint" and category columns,}
}
\value{
dataframe transformed for plotting with category_plot
}
\description{
Transform rate data to rate plot data.
}
