% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_plot_data.r
\name{category_plot_table_data}
\alias{category_plot_table_data}
\title{Category Plot Table Data}
\usage{
category_plot_table_data(plot_data)
}
\arguments{
\item{plot_data}{dataframe with "identifier", "timepoint" and category columns,}
}
\value{
dataframe transformed for printing in an individual report
}
\description{
Transform category plot data for printing in a report table
}
