% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.r
\name{check_input}
\alias{check_input}
\alias{check_col_names}
\alias{check_col_count}
\title{Check Input}
\usage{
check_input(df, expected_colnames)

check_col_names(df, expected_colnames)

check_col_count(df, expected_colnames)
}
\arguments{
\item{df}{dataframe of input data to verify}

\item{expected_colnames}{vector of names that need to match the column headers}
}
\description{
Make a pass through data for conformance to expected col names and classes.
}
\section{Functions}{
\itemize{
\item \code{check_col_names}: Check observed column names in df against expected.

\item \code{check_col_count}: Check observed number of columns in df against expected.
}}

