% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespan_text.r
\name{date_axis_labeler}
\alias{date_axis_labeler}
\title{Date axis labeler}
\usage{
date_axis_labeler(timepoints)
}
\arguments{
\item{timepoints}{character vector of all in form of "2018 Jan"}
}
\description{
Utility function to convert timepoints factor into string for title.
}
