% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_recent_times.r
\name{backtimes_from_latest}
\alias{backtimes_from_latest}
\title{Back Times from Latest}
\usage{
backtimes_from_latest(df, x)
}
\arguments{
\item{df}{input data frame with fy, quart columns}
}
\value{
dataframe with columns fy and quart that lists x timepoints including the max from df
}
\description{
Find most recent fy and quart, then count backward x quarters
}
\section{Functions}{
\itemize{
\item \code{backtimes_from_latest}: Recent Times
}}

