% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_clc_data.r
\name{filter_clc_data}
\alias{filter_clc_data}
\title{Filter CLC Data}
\usage{
filter_clc_data(df)
}
\arguments{
\item{df}{clc dataframe}
}
\value{
dataframe of filtered CLC data
}
\description{
Handle the special cases of CLC data
}
