% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dementia_data.r
\name{filter_dementia_data}
\alias{filter_dementia_data}
\title{Filter Dementia Data}
\usage{
filter_dementia_data(df)
}
\arguments{
\item{df}{dementia dataframe}
}
\value{
dataframe of filtered Dementia data
}
\description{
Handle the special cases of Dementia data
}
