% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_hbpc_data.r
\name{filter_hbpc_data}
\alias{filter_hbpc_data}
\title{Filter HBPC Data}
\usage{
filter_hbpc_data(df)
}
\arguments{
\item{df}{hbpc dataframe}
}
\value{
dataframe of filtered HBPC data
}
\description{
Handle the special cases of HBPC data
}
\note{
Rename id column for consistency with clc then filter
}
