% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_all_tex.r
\name{generate_report}
\alias{generate_report}
\title{Generate Report}
\usage{
generate_report(id, rate_df, category_df, template_path, build_dir, reports_dir)
}
\description{
Generate the report for an individual id.
}
