% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_print_lim.r
\name{lower_print_lim}
\alias{lower_print_lim}
\title{Lower Print Lim}
\usage{
lower_print_lim(x)
}
\arguments{
\item{x}{vector of counts}
}
\description{
Hueristic for determining the lowest number character to print on a bar.
}
