% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.r
\name{process_data}
\alias{process_data}
\title{Process Data}
\usage{
process_data(df, envir)
}
\arguments{
\item{df}{dataframe of CLC or HBPC data}

\item{envir}{Environment with constants for processing df.}
}
\value{
list of dataframes: rate and category data frames.
}
\description{
Process input data frame of clc or hbpc data
}
