% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_plot.r
\name{rate_extra_colnames}
\alias{rate_extra_colnames}
\title{Rate Extra Colnames}
\usage{
rate_extra_colnames(cnames)
}
\description{
set diff of column names in plot data with expected names
}
\section{Functions}{
\itemize{
\item \code{rate_extra_colnames}: Plot
}}

