% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_plot.r
\name{rate_plot}
\alias{rate_plot}
\title{Rate Plot}
\usage{
rate_plot(plot_data, plot_title = "", y_label = "", line_label = "",
  stack_labels = c(""))
}
\arguments{
\item{plot_data}{rate plot dataframe}

\item{plot_title}{string for plot title}

\item{y_label}{string for y-axis label}

\item{line_label}{legend text for line that represents total admissions}

\item{stack_labels}{legend text for the numerator and misses values.}
}
\description{
Create a rate plot figure.
}
