% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_plot_data.r
\name{rate_plot_data}
\alias{rate_plot_data}
\title{Rate Plot Data}
\usage{
rate_plot_data(input_data)
}
\arguments{
\item{input_data}{dataframe with "identifier", "numerator" and "misses" columns,}
}
\value{
dataframe transformed for plotting with rate_plot
}
\description{
Transform rate data to rate plot data.
}
