% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.r
\name{read_data}
\alias{read_data}
\alias{read_clc_data}
\alias{read_hbpc_data}
\alias{read_dementia_data}
\title{Read Data}
\usage{
read_data(path, classes)

read_clc_data(path)

read_hbpc_data(path)

read_dementia_data(path)
}
\arguments{
\item{path}{path to clc or hbpc csv file}

\item{classes}{vector of classes passed to colClasses}
}
\value{
dataframe or NULL
}
\section{Functions}{
\itemize{
\item \code{read_clc_data}: Convenience function for reading CLC data

\item \code{read_hbpc_data}: Convenience function for reading HBPC data

\item \code{read_dementia_data}: Convenience function for reading dementia data
}}

