% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_all.r
\name{report_all}
\alias{report_all}
\title{Report All}
\usage{
report_all(df_list, envir, config, output_dir)
}
\arguments{
\item{df_list}{list of data frames.  One for rate and one for category}

\item{envir}{environment corresponding to the report template and data provided.}

\item{config}{external configuration}

\item{output_dir}{destination for reports}
}
\description{
Run report for all IDs in the data.
}
