% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_all.r
\name{report_one}
\alias{report_one}
\title{Report One}
\usage{
report_one(id, df_list, envir, config, output_dir)
}
\arguments{
\item{id}{by which to subset}

\item{envir}{with name matching the convention for the corresponding template}

\item{config}{external configuration information}

\item{output_directory}{path to directory where output file will be deposited.}
}
\description{
Create report for a single id
}
