% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespan_text.r
\name{timespan_text}
\alias{timespan_text}
\title{Timespan Text}
\usage{
timespan_text(timepoints)
}
\arguments{
\item{timepoints}{character vector of all in form of "2018\\nQ3"}
}
\description{
Utility function to convert timepoints factor into string for title.
}
