% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_category_sums.r
\name{calc_category_sums}
\alias{calc_category_sums}
\title{Calculate Category Sums}
\usage{
calc_category_sums(flt_data, id_cols, numer_categories)
}
\arguments{
\item{flt_data}{dataframe of filtered data}

\item{id_cols}{names of columns to use as id}

\item{numer_categories}{names of columns grouped by category}
}
\value{
dataframe of category sums
}
