% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_recent_times.r
\name{filter_recent_times}
\alias{filter_recent_times}
\title{Filter Recent Times}
\usage{
filter_recent_times(df, x)
}
\arguments{
\item{df}{the hbpc or clc data frame: has fy, quart, sta6a}

\item{x}{number of most recent time points to include}
}
\value{
dataframe of filtered data
}
\description{
Trim data to most recent x timepoints and fill in rows for missing times
}
