% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{main}
\alias{main}
\title{Main}
\usage{
main(input_dir, config_path, output_dir = NULL)
}
\arguments{
\item{input_dir}{path to directory that contains clc.csv or hbpc.csv}

\item{config_path}{path to configuration file}

\item{output_dir}{path to output directory. Defaults to input_dir}
}
\value{
boolean TRUE indicates successful run.
}
\description{
Build all reports for input data in input directory
}
\note{
This is the entry point for the application.
}
